fix: resolve circular import chain and missing dependencies
- Extract ToolSpec and ToolCategory to separate tool_specs.py module
- Fix double src import paths throughout codebase
- Use TYPE_CHECKING and runtime imports to break agent circular dependencies
- Add missing dependencies: trafilatura, gradio, limits, python-dateutil
- Correct analytics module import paths
- Add Union import to app.py
- Create configs/__init__.py for Hydra
- Update graph instantiation to include workflow nodes
