+++ DeepResearch/src/agents/workflow_orchestrator.py
@@ -17,18 +17,18 @@ from omegaconf import DictConfig
 from pydantic_ai import Agent, RunContext
 from pydantic import BaseModel, Field
 
-from ..src.datatypes.workflow_orchestration import (
+from ..datatypes.workflow_orchestration import (
     WorkflowOrchestrationConfig, WorkflowExecution, WorkflowResult, WorkflowStatus,
     WorkflowType, AgentRole, DataLoaderType, WorkflowComposition, OrchestrationState,
     HypothesisDataset, HypothesisTestingEnvironment, ReasoningResult
 )
-from ..src.datatypes.rag import RAGConfig, RAGResponse, BioinformaticsRAGResponse
-from ..src.datatypes.bioinformatics import FusedDataset, ReasoningTask, DataFusionRequest
+from ..datatypes.rag import RAGConfig, RAGResponse, BioinformaticsRAGResponse
+from ..datatypes.bioinformatics import FusedDataset, ReasoningTask, DataFusionRequest
 
 if TYPE_CHECKING:
-    from ..src.agents.bioinformatics_agents import AgentOrchestrator
-    from ..src.agents.search_agent import SearchAgent
-    from ..src.agents.research_agent import ResearchAgent
+    from ..agents.bioinformatics_agents import AgentOrchestrator
+    from ..agents.search_agent import SearchAgent
+    from ..agents.research_agent import ResearchAgent
 
 
 class OrchestratorDependencies(BaseModel):
