+++ DeepResearch/src/statemachines/search_workflow.py
@@ -13,9 +13,9 @@ from pydantic_graph import Graph, Node, End
 from ..tools.websearch_tools import WebSearchTool, ChunkedSearchTool
 from ..tools.analytics_tools import RecordRequestTool, GetAnalyticsDataTool
 from ..tools.integrated_search_tools import IntegratedSearchTool, RAGSearchTool
-from ..src.datatypes.rag import Document, Chunk, RAGQuery, RAGResponse
-from ..src.utils.execution_status import ExecutionStatus
-from ..src.utils.execution_history import ExecutionHistory, ExecutionItem
+from ..datatypes.rag import Document, Chunk, RAGQuery, RAGResponse
+from ..utils.execution_status import ExecutionStatus
+from ..utils.execution_history import ExecutionHistory, ExecutionItem
 from ...agents import SearchAgent, AgentDependencies, AgentResult, AgentType
