+++ DeepResearch/tools/analytics_tools.py
@@ -12,7 +12,7 @@ from pydantic import BaseModel, Field
 from pydantic_ai import Agent, RunContext
 
 from .base import ToolSpec, ToolRunner, ExecutionResult
-from .analytics import record_request, last_n_days_df, last_n_days_avg_time_df
+from ..src.utils.analytics import record_request, last_n_days_df, last_n_days_avg_time_df
 
 
 class AnalyticsRequest(BaseModel):
