+++ DeepResearch/tools/websearch_cleaned.py
@@ -11,7 +11,7 @@ from dateutil import parser as dateparser
 from limits import parse
 from limits.aio.storage import MemoryStorage
 from limits.aio.strategies import MovingWindowRateLimiter
-from analytics import record_request, last_n_days_df, last_n_days_avg_time_df
+from ..src.utils.analytics import record_request, last_n_days_df, last_n_days_avg_time_df
 
 # Configuration
 SERPER_API_KEY_ENV = os.getenv("SERPER_API_KEY")
