+++ pyproject.toml
@@ -10,11 +10,15 @@ authors = [
 ]
 dependencies = [
   "beautifulsoup4>=4.14.2",
+  "gradio>=5.47.2",
   "hydra-core>=1.3.2",
+  "limits>=5.6.0",
   "pydantic>=2.7",
   "pydantic-ai>=0.0.16",
   "pydantic-graph>=0.2.0",
+  "python-dateutil>=2.9.0.post0",
   "testcontainers>=4.8.0",
+  "trafilatura>=2.0.0",
 ]
 
 [project.optional-dependencies]
